+++ src/extensions/agent_modules/cloud_communication/eg_mqtt_broker_client/CMakeLists.txt
@@ -9,7 +9,9 @@ find_package (azure_c_shared_utility REQUIRED)
 find_package (Parson REQUIRED)
 
 target_sources (${target_name} PRIVATE src/adu_agent_info_management.c
+                                       src/adu_agent_info_management_internal.c
                                        src/adu_enrollment_management.c
+                                       src/adu_enrollment_management_internal.c
                                        src/adu_mqtt_client_module.c
                                        src/adu_updates_management.c
                                        src/agentinfo_request_operation.c
